Jaccard distance: Move from a fallback to its own class
